% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive.R
\name{archive_package}
\alias{archive_package}
\title{Archive package in CRAN}
\usage{
archive_package(cran_root, package_name)
}
\arguments{
\item{cran_root}{The folder containing the CRAN.}

\item{package_name}{The name of the package \strong{without} the version & file extension.}
}
\description{
If there are multiple \code{tar.gz}/\code{zip} files for the same package, all but the one with the highest
version are moved/deleted.
}
