% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_files.R
\name{archive_package_files}
\alias{archive_package_files}
\title{Files in archive}
\usage{
archive_package_files(cran_root)
}
\arguments{
\item{cran_root}{The folder containing the CRAN.}
}
\description{
List the files in the \code{Archive} folder of a local CRAN \emph{sans} the folder.
}
