% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{archive_path}
\alias{archive_path}
\title{The path of the Archive folder}
\usage{
archive_path(cran_root)
}
\arguments{
\item{cran_root}{The folder containing the CRAN.}
}
\value{
The path of \code{archive.rds} relative to \code{cran_root}
}
\description{
The path of the Archive folder
}
