% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive.R
\name{archive_single_binary_package}
\alias{archive_single_binary_package}
\title{Archive outdated versions of a package}
\usage{
archive_single_binary_package(package_files)
}
\arguments{
\item{package_files}{A vector with filenames.}
}
\description{
{cranitor} only retains the most recent version of a binary package; older versions are deleted.
If there are multiple elements in \code{package_files} all but the most recent are deleted.
}
