% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive.R
\name{archive_source_package}
\alias{archive_source_package}
\title{Archive source package}
\usage{
archive_source_package(cran_root, source_files)
}
\arguments{
\item{cran_root}{The folder containing the CRAN.}

\item{source_files}{The \code{tar.gz} filenames with source code \emph{sans} path.}
}
\description{
Move \code{tar.gz} files with source code to the correct location in the CRAN.
}
