% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive.R
\name{archive_windows_package}
\alias{archive_windows_package}
\title{Archive Windows package}
\usage{
archive_windows_package(cran_root, bin_files)
}
\arguments{
\item{cran_root}{The folder containing the CRAN.}

\item{bin_files}{The \code{zip} files with compiled code.}
}
\description{
Move \code{zip} files with compiled source code to the correct location in the CRAN.
}
