% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive.R
\name{basename_from_targz}
\alias{basename_from_targz}
\title{Extract package name}
\usage{
basename_from_targz(targz)
}
\arguments{
\item{targz}{A vector of filenames.}
}
\value{
A vector of package names.
}
\description{
A packed source code file is of the form \verb{<package name>_<version>.tar.gz}. This function
returns \verb{<package name>} from such a file name.
}
