% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{build_update_cran}
\alias{build_update_cran}
\title{Build package files and import to CRAN}
\usage{
build_update_cran(cran_root, ...)
}
\arguments{
\item{cran_root}{The folder containing the CRAN.}

\item{...}{Input for \code{\link[devtools:build]{devtools::build()}} besides \code{binary}}
}
\description{
Update DESCRIPTION, build package files for CRAN and import them.
}
