% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean_cran}
\alias{clean_cran}
\title{Clean a CRAN}
\usage{
clean_cran(cran_root, list = FALSE)
}
\arguments{
\item{cran_root}{The folder containing the CRAN.}

\item{list}{\verb{[logical]} List unnecessary files or delete them.}
}
\description{
Clean a CRAN by archiving as needed for each platform and removing unnecessary files.
}
