% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cranitor-package.R
\docType{package}
\name{cranitor}
\alias{cranitor}
\alias{cranitor-package}
\title{\code{cranitor} package}
\description{
Taking care of a CRAN from initialization to updating metadata when new packages/package versions are included.
}
\details{
See the README on \href{https://github.com/robertdj/cranitor}{GitHub}.
}
\author{
\strong{Maintainer}: Robert Dahl Jacobsen \email{git@dahl-jacobsen.dk}

}
