% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.R
\name{create_empty_package}
\alias{create_empty_package}
\title{Make an empty package}
\usage{
create_empty_package(package_name, version, ...)
}
\arguments{
\item{package_name}{The name of the package}

\item{version}{The version of the package}

\item{...}{Arguments for \code{\link[pkgbuild:build]{pkgbuild::build()}}}
}
\value{
The path of the built package.
}
\description{
Make an empty package
}
\details{
The package consists of a \code{DESCRIPTION} file and a \code{NAMESPACE} file.
}
