% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{formatted_archive_metadata}
\alias{formatted_archive_metadata}
\title{List the archived packages}
\usage{
formatted_archive_metadata(cran_root)
}
\arguments{
\item{cran_root}{The folder containing the CRAN.}
}
\description{
List the archived packages' metadata in the form expected for a CRAN.
}
