% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_source_package}
\alias{import_source_package}
\title{Import source package}
\usage{
import_source_package(cran_root, package)
}
\arguments{
\item{cran_root}{The folder containing the CRAN.}

\item{package}{A vector with filenames (including path) of the \code{tar.gz} files.}
}
\description{
Import source package \code{tar.gz} file into a local CRAN by copying the file into the
appropriate folder.
}
