% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_win_package}
\alias{import_win_package}
\title{Import Windows package}
\usage{
import_win_package(cran_root, package, r_version)
}
\arguments{
\item{cran_root}{The folder containing the CRAN.}

\item{package}{A vector with filenames (including path) of the \code{zip} files.}
}
\description{
Import Windos package \code{zip} file into a local CRAN by copying the file into the
appropriate folder.
}
