% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{is_valid_linux_dir}
\alias{is_valid_linux_dir}
\title{Is folder a valid Linux dir}
\usage{
is_valid_linux_dir(linux_dir)
}
\arguments{
\item{linux_dir}{Folder to test}
}
\value{
\code{TRUE} or \code{FALSE}. If \code{FALSE} as message is also written.
}
\description{
A folder is considered valid if it contains "\strong{linux}" and a 3 digit R version
}
