% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_files.R
\name{list_package_files}
\alias{list_package_files}
\title{List packages}
\usage{
list_package_files(package_root, package_name)
}
\arguments{
\item{package_root}{The folder with the \code{tar.gz}/\code{zip} files.}

\item{package_name}{Name of the package.}
}
\value{
A vector with file names.
}
\description{
List \code{tar.gz} or \code{zip} archives of a package sorted according to modification time with newest
first.
}
