% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive.R
\name{make_archive_metadata}
\alias{make_archive_metadata}
\title{Save archive metadata}
\usage{
make_archive_metadata(cran_root)
}
\arguments{
\item{cran_root}{The folder containing the CRAN.}
}
\description{
Save archive metadata to \code{src/contrib/Meta/archive.rds}.
}
