% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.R
\name{make_demo_cran}
\alias{make_demo_cran}
\title{Make a demo CRAN}
\usage{
make_demo_cran(cran_root = NULL)
}
\arguments{
\item{cran_root}{The folder containing the demo CRAN. If \code{NULL}, the folder \code{demo_cran} will be
created in a temporary folder.}
}
\value{
The folder \code{cran_root}
}
\description{
Make a demo CRAN
}
