% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{make_local_cran}
\alias{make_local_cran}
\title{Make a local CRAN}
\usage{
make_local_cran(cran_root)
}
\arguments{
\item{cran_root}{The folder containing the CRAN.}
}
\description{
Make a local CRAN with source packages and binary packages for Windows.
}
