% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{package_metadata}
\alias{package_metadata}
\title{Get package metadata}
\usage{
package_metadata(package_files)
}
\arguments{
\item{package_files}{A vector of strings with locations of \code{tar.gz} source files.}
}
\value{
A \code{data.frame} with columns \code{size}, \code{isdir}, \code{mode}, \code{mtime}, \code{ctime}, \code{atime}, \code{uid},
\code{gid}, \code{uname} and \code{grname}. This is \code{\link[=file.info]{file.info()}} for the package file.
}
\description{
Get the metadata of a package from its \code{tar.gz} source file.
}
