% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{source_package_dir}
\alias{source_package_dir}
\title{The path of the source package folder}
\usage{
source_package_dir(cran_root)
}
\arguments{
\item{cran_root}{The folder containing the CRAN.}
}
\description{
The path of the source package folder
}
