% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{testdata_path}
\alias{testdata_path}
\title{Get test data}
\usage{
testdata_path(filename)
}
\arguments{
\item{filename}{The filename of the test file as it appears in \code{inst/testdata} \emph{without} the
path.}
}
\value{
The full path of \code{filename}.
}
\description{
Get test data in the \emph{installed} \code{cranitor} package. Used in the tests of \code{cranitor}.
}
