% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{win_package_dir}
\alias{win_package_dir}
\title{The path of the Windows package folder}
\usage{
win_package_dir(cran_root, r_version = getRversion())
}
\arguments{
\item{cran_root}{The folder containing the CRAN.}

\item{r_version}{An \code{R_system_version} -- the result of e.g. \code{\link[=getRversion]{getRversion()}}.}
}
\description{
The path of the Windows package folder
}
