% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update_cran}
\alias{update_cran}
\title{Update local CRAN with the package}
\usage{
update_cran(cran_root, targz_file, zip_file = NULL)
}
\arguments{
\item{cran_root}{The folder containing the CRAN.}

\item{targz_file}{The location of the \code{tar.gz} file for source.}

\item{zip_file}{The location of the \code{zip} file for Windows.}
}
\description{
Import package files into a local CRAN and update the metadata. Check the README in the repo.
}
