% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{win_package_dir}
\alias{win_package_dir}
\title{The path of the Windows package folder}
\usage{
win_package_dir(cran_root)
}
\arguments{
\item{cran_root}{The folder containing the CRAN.}
}
\description{
The path of the Windows package folder
}
